FIX: Default CID is unusable
